optimize simulator
I should note that while this *technically* optimizes the actual app as well, but the Swift compiler for on-device will notice isSimulator will never be set, so compiler optimizations will remove it and, in the end, produce the same code as this optimization. However, since it is set on Simulator, it won't optimize it, meaning this produces slower code. I know Simulator optimizations don't matter to an average user, but this change (slightly) benefits development (and even if it won't matter for the on-device app, doing this optimization in the code itself rather than relying on compiler optimizations to do it for you is a better practice :P).