fix: various optimizations (#56)
* optimize simulator

I should note that while this *technically* optimizes the actual app as well, but the Swift compiler for on-device will notice isSimulator will never be set, so compiler optimizations will remove it and, in the end, produce the same code as this optimization. However, since it is set on Simulator, it won't optimize it, meaning this produces slower code. I know Simulator optimizations don't matter to an average user, but this change (slightly) benefits development (and even if it won't matter for the on-device app, doing this optimization in the code itself rather than relying on compiler optimizations to do it for you is a better practice :P).

* optimize simulator

I should note that while this *technically* optimizes the actual app as well, but the Swift compiler for on-device will notice isSimulator will never be set, so compiler optimizations will remove it and, in the end, produce the same code as this optimization. However, since it is set on Simulator, it won't optimize it, meaning this produces slower code. I know Simulator optimizations don't matter to an average user, but this change (slightly) benefits development (and even if it won't matter for the on-device app, doing this optimization in the code itself rather than relying on compiler optimizations to do it for you is a better practice :P).

* remove IOKit

At least from what I can see, there is no place where the loader uses anything from it(?). I tried compiling without it, and everything seems fine and doesn't look to have any effect on the compiled binary.

Removing this makes development experience much better, as before you couldn't compile this on systems without this IOKit header, now with the changes it's possible.

* minor optimization

* minor optimization: we already know it's rootless

* optimize JsonVC

Improved logic. No longer preforms a lookup table when section is 0. When section is 1, instead of doing the lookup table we replace it entirely with if statements since two comparisons are faster than a lookup table.

* more optimization

* comment changes for clarity

* spawn is always called as root

since spawn is always called as root, we can remove this arg

* spawn is always called as root

since spawn is always called as root, we can remove this arg

* spawn is always called as root

since spawn is always called as root, we can remove this arg

* spawn is always called as root

since spawn is always called as root, we can remove this arg

* spawn is always called as root

since spawn is always called as root, we can remove this arg

* remove systemVersion from envInfo

this is a slight optimization

* remove systemVersion from envInfo

this is a slight optimization

* very minor optimization: just use envInfo.jsonURI itself